Edit Makefile

+++ src/ur.h
@@ -2,7 +2,7 @@
 #define H_URAV
 
 
-double urav (double a, double b, double c, double *x1, double *x2);
+double urav (int a, int b, int c, double *x1, double *x2);
 
 #endif
